+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **[Breaking change]** Add `Error` raw action to represent invalid actions.
 - **[Breaking change]** Add `End` raw action.
+- **[Breaking change]** Update `Cfg` representation to ensure it is non-empty.
 - **[Feature]** Add error information to `CfgErrorBlock`.
 
 ## Typescript